requirements.txt
ultralytics>=8.3.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
PyYAML>=6.0
numpy>=1.24.0,<2.0
tqdm>=4.65.0
requests>=2.31.0
